requirements.txt
google-ads>=25.0.0
mcp>=1.1.0
httpx>=0.27.0
pydantic>=2.0.0
google-auth-oauthlib>=1.0.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
cachetools>=5.3.0
redis>=5.0.0

# Export and reporting dependencies
openpyxl>=3.1.2
reportlab>=4.0.0
matplotlib>=3.8.0
scipy>=1.11.0
statsmodels>=0.14.0
